Correcciones a las horas de prácticas
Correcciones a las horas de prácticas

+++ PresentationLayer/ConsultProgress.xaml.cs
@@ -14,7 +14,7 @@ namespace PresentationLayer
 	{
 		private readonly string _enrollment;
 		private int _hourCovered;
-		public int ProfessionalPracticesState { get; set; }
+		public float ProfessionalPracticesState { get; set; }
 
 		public ConsultProgress(string practicionerEnrollment)
 		{
@@ -88,7 +88,7 @@ namespace PresentationLayer
 
 		private void CalculateProgress()
 		{
-			int quantity = _hourCovered * 100 / 480;
+			float quantity = _hourCovered * 100 / 480;
 			labelPercentage.Content = quantity + " % " +_hourCovered+" horas cubiertas de 480 horas";
 			ProfessionalPracticesState = quantity ;
 		}
